+++ healthper/src/main/java/umc/healthper/repository/PostRepository.java
@@ -26,7 +26,7 @@ public class PostRepository {
     /**
      * 게시글 조회
      */
-    public Post findById(int id) {
+    public Post findById(Long id) {
         return em.find(Post.class, id);
     }
 
@@ -40,7 +40,7 @@ public class PostRepository {
     /**
      * 게시글 삭제
      */
-    public void removeById(int id) {
+    public void removeById(Long id) {
         Post findPost = em.find(Post.class, id);
         // 댓글들도 삭제해야 하는가?
         // -> 아니다. 댓글정보 클릭해서 게시글 조회하려고 할 때는 "삭제된 게시글입니다"만 띄워주면 될 것
